refactor: rename UpdateFetchers.ts to update-fetchers.ts
