+++ src/lib/components/composites/settings/project-settings/slr/FetcherRemovalDialog.svelte
@@ -3,7 +3,7 @@
     import ActionErrorAlert from "$lib/components/composites/utils/ActionErrorAlert.svelte";
     import type { ActionError } from "$lib/model/action-error";
     import { Project, Project_Settings } from "$lib/model/api/project";
-    import { updateFetchers } from "./UpdateFetchers";
+    import { updateFetchers } from "./update-fetchers";
 
     interface Props {
         projectId: string;
